Remove Responder and change how to require model dependency
