--- lib/trailblazer/operation/responder.rb
@@ -1,8 +1,0 @@
-class Trailblazer::Operation
-  module Responder
-    def errors
-      return [] if valid?
-      [1]
-    end
-  end
-end
