+++ test/fake_app/song/operations.rb
@@ -3,7 +3,6 @@ require 'trailblazer/autoloading'
 class Song < ActiveRecord::Base
   class Create < Trailblazer::Operation
     include Model
-    include Responder
 
     model Song, :create
 
@@ -39,7 +38,7 @@ end
 
 class Band < ActiveRecord::Base
   class Create < Trailblazer::Operation
-    include Model, Responder#, Representer
+    include Model#, Representer
     model Band, :create
 
     contract do
